Check domain when looking up ONNX op deserializer
Use (domain, op_type) as the key for ONNX op deserializer functions instead of
just op_type. When reading `NodeProto` messages, deserialize the `domain` field
and use it as part of the key, defaulting to "ai.onnx" if not present.

This avoids intepreting operators incorrectly if they have an op_type that
matches a built-in operator but use a custom domain. The "com.microsoft" domain
has several ops with names that overlap with ops from the "ai.onnx" domain [^1].

This will also make it possible to support non-standard ONNX operators. See
https://github.com/robertknight/rten/issues/578.

[^1]: https://github.com/microsoft/onnxruntime/blob/main/docs/ContribOperators.md