Support 4-bit weight-only quantization in rten-gemm
Support a new type of RHS input in rten-gemm where the LHS input is f32 and the
RHS input is quantized to 4-bits in blocks along the K dimension. The 4-bit
weights are dequantized during the existing packing stage.

The input format aligns with the `MatMulNBits` contrib operator described at
https://github.com/microsoft/onnxruntime/blob/main/docs/ContribOperators.md#com.microsoft.MatMulNBits.

The initial implementation only supports the default zero point of 2^(n_bits -
1) (ie. 8 for 4-bit elements).

Part of https://github.com/robertknight/rten/issues/578.