Revise ort-quantize.py interface and support QDQ for nbits quantizer
Replace the `--mode` flag with a subcommand which selects the quantizer to use.
This approach more cleanly maps to how the implementation works and allows the
selected mode to determine available flags. Use this to add a `--qdq` flag for
the nbits quantizer only, which causes the quantizer to use DequantizeLinear
plus MatMul rather than MatMulNBits.