Support `ignore_merges` option for BPE tokenizers
Support the `ignore_merges` option that was introduced in the tokenizers library
in https://github.com/huggingface/tokenizers/pull/1493. This causes string
pieces being encoded (after pre-tokenization) to be matched against the
vocabulary before applying merge rules. For Llama 3 this affects a very small
number of tokens (<0.5%). The downside of enabling this option is that it
requires extra memory/copies to keep the vocab around, so we only do this when
needed.

Fixes https://github.com/robertknight/rten/issues/453.