Make `Bpe::merges` keys token IDs rather than ranks
Change the `Bpe::merges` field so that keys are pairs of token IDs rather than
ranks and the value is a `(rank, id)` pair rather than just a rank. This is a
more natural way to represent the merge list which simplifies other parts of the
Bpe model. It also aligns with the corresponding `MergeMap` type in the
`tokenizers` crate and will make it possible to implement the `ignore_merges`
option that is used by the Llama 3 tokenizer.

Part of https://github.com/robertknight/rten/issues/453.