Use VNNI in AVX-512 GEMV kernel if available
Use VNNI if available for vector-matrix products in int8 AVX-512 matmul kernel.
This is faster and fixes a test failure on systems with VNNI enabled. The test
failure happened because the Kernel reported that it did not suffer from
saturation if VNNI was enabled, and so the test generated inputs using the full
int8 range. However if the input had M=1, the GEMV code path would be used which
didn't use VNNI and so did encounter saturation.

Note that for the non-GEMV path, VNNI was already used.