Group inputs and outputs with similar names in CLI output
Group inputs and outputs whose names are the same except for a number somewhere
in the name, eg. "present.0.key" and "present.1.key". Inputs are only grouped if
their other metadata (dtype, shape) is the same.

LLM models tend to have a large number of such inputs because there are
key-value cache inputs and outputs for each layer. This made for a lot of
scrolling when reading CLI output. With grouping, the summary of model inputs
for Llama 3 3B looks like:

```
Inputs
  input_ids: i32 [batch_size, sequence_length]
  attention_mask: i32 [batch_size, past_sequence_length + sequence_length]
  position_ids: i32 [batch_size, sequence_length]
  past_key_values.{0-27}.key: f32 [batch_size, 8, past_sequence_length, 128]
  past_key_values.{0-27}.value: f32 [batch_size, 8, past_sequence_length, 128]
```