Update Layout trait docs to mention that tensors implement it
The frontmatter of the rten-tensor create mentioned that tensors implement the
Layout trait, but the documentation for the trait itself did not. This is
important since any methods added to layouts become available on tensors as
well. This has consequences for which methods can be added to the trait: Methods
names which conflict with other traits (eg. AsView) implemented by tensor types
should be avoided, as well as methods which make sense for standalone layout
types like NdLayout, but not tensor types (eg. `squeezed`).