Support dynamic quantization of LHS in block-quantized matmul on Arm
On Arm, support quantizing the LHS to 8-bit ints (Q8) in block-quantized matmul.
This enables the use of int8 dot product instructions. Other architectures
continue to use an un-quantized LHS. Quantization is enabled if the
`accuracy_level` attribute for MatMulNBits is set to 4 or higher.

On Llama 3.2 3B on an M3 Pro this improves generation speed from 70ms/token to
28ms/token. For comparison, ORT and llama.cpp both achieve around
~21-23ms/token.