Support int8 compute in MatMulNBits on x86
Add AVX2 and AVX512-VNNI implementations of the int8 dot product SIMD ISAs and
use these to support x86 in GEMV with Q4 RHS and Q8 LHS. On x86 the LHS input to
the int8 dot product instructions is unsigned. This is handled by skipping the
subtraction of the zero point from the RHS input to keep it unsigned, doing
`dot(col, row)` (note the LHS and RHS are swapped) and then subtracting `lhs_sum
* rhs_zero_point` afterwards to compensate. Under AVX2 the dot product
instruction sequence can encounter saturation when adding intermediate i16
values. This can't happen in our case however because the LHS input to the dot
product is in the Q4 range (ie. 0-15).