Implement GlobalMaxPool operator
Replace the GlobalAveragePool implementation with a general global pooling
function which is parametrized by a kernel and implement both GlobalAveragePool
and GlobalMaxPool using this. The new global pooling operators follow the same
implementation design as the reduction ops, where the input is divided into
contiguous chunks to which a kernel is applied.

In addition to better efficiency, the new implementation also supports all
inputs with at least 2 dimensions, instead of just NCHW tensors.