Support external data when loading models from byte buffers
Add a `ModelOptions::external_data` method which allows providing the content of
an external file when loading a model from a byte buffer, via
`ModelOptions::load`. This enables loading ONNX models with external data in
contexts where a file system is not available (eg.  in a browser).

If `ModelOptions::external_data` is used but the model is loaded via
`ModelOptions::load_file`, the in-memory external data is ignored.

The API could have alternatively been a new optional argument to `Model::load`,
although this would be a breaking change.