Omit braces when input group has one number
When showing groups of inputs or outputs in CLI output, omit braces around group
names if there is only one node in the group. If for example there are two
inputs "inputs.0" and "inputs.1" with different shapes, these should be put into
two separate groups and their names shown as "inputs.0" and "inputs.1" instead
of "inputs.{0}" and "inputs.{1}".