Support SmolLM v3 in the Qwen chat example
Make a minor change to the Qwen chat example so that it works with other models
that use the same prompt format. So far only SmolLM v3 has been tested.

This is only "basic" support in the sense that there isn't any special handling
of thinking tokens or other new capabilities of this model.