Fix panic during sampling if logits contains a NaN
Fix panic in multinomial sampler if the input logits contain a NaN. In this case
the softmax operation would spread the NaN to all other logits and the
`multinomial` helper would return `None`. This also happened if the logits
contained a positive (but not negative) infinity.

This issue could be handled by making the generator return an error, or it could
be handled as in this commit by returning a default token ID of zero. I chose
the latter option here only because it is simpler to implement.