Support variable number of batch dims in MatMulNBits
Support A inputs with 2+ dimensions in MatMulNBits, instead of requiring
the rank to be exactly three. Internally the LHS input is reshaped to
`[batch_len * M, K]` before the multiplication, similar to how the
`MatMul` implementation works.

Also always pass the RHS input to the GEMM implementation as a single matrix,
instead of duplicating it. This avoids unnecessary repacking of the RHS matrix
for each LHS matrix in the batch.