Fix rten-text tests when run standalone
This is a follow up to https://github.com/robertknight/rten/pull/1095
which updates some imports that were missed in that PR.

This issue was not found by CI because running tests for all workspace crates
via `cargo test --workspace` succeeded, but running tests for rten-text alone
using `cargo test -p rten-text` failed. This is presumably due to different
features being activated for the serde crate.