Make zero-copy deserialization of BPE merges work as intended
https://github.com/robertknight/rten/pull/1099 made a change to tokenizer
deserialization intended to avoid string allocations for most entries in the BPE
merge list. While performance did improve for the whole change set, the change
to the `MergeList` type didn't actually work as intended. See the
users.rust-lang.org post mentioned in this commit. Fix this by using a newtype
which does work as expected. With this only ~0.7% of the strings in the merge
list require allocations.

This reduces JSON parse times from ~62ms to ~54ms (-12%) for the Llama 3
tokenizer.