Support Pow with i32 base and i32 exponent
Extend the Pow operator to support i32 base and i32 exponents. Previously i32
base was supported only with an f32 exponent.

In implementing this, some differences between rten, ONNX Runtime and
PyTorch/numpy were uncovered. Update the behavior to align with PyTorch and
NumPy, except for the case where the base and exponent are integral types and
the exponent is negative. This was done to avoid making the operation fallible.