Add ComputeShape operator and fusion
Add a fusion which replaces `Shape` operators with a `ComputeShape` operator
that produces the same output using information from shape inference. This
information includes a mix of static values and dynamic values extracted from
graph inputs.

The main benefit of this fusion is to remove a consumer of the operator output
which feeds into the Shape op. This frees up that operator to be included in
fusions. In the context of Llama 3, this will enable fusing the RepeatInterleave
operator that broadcasts keys with the query-key matmul.

A hazard with this change is that it relies on metadata baked into the model
from shape inference to be correct. This could eventually be avoided by
implementing shape inference analysis within RTen.
Part of https://github.com/robertknight/rten/issues/1091.