Add built-in shape and type inference
Add a built-in system to infer the shapes and types of operator outputs. When
enabled, this runs before graph optimizations. This generates metadata that can
be used by the optimizer to verify that conditions that a particular fusion has
on operator inputs and outputs (eg. on type, rank, dimension sizes) are
satisfied.

ONNX already provides Python-based tools for running shape inference and
embedding the resulting metadata in graphs. The advantages of RTen having a
built-in implementation are:

 - It works even if shape inference has not been run on the model

 - It works for non-standard operators (eg. MatMulNBits)

 - In some cases it can infer shapes and values which the built-in ONNX tools
   are not able to infer.

To begin with, this system is opt-in via `ModelOptions::enable_shape_inference`.
The intent is to enable it by default when it is more mature.

The components of the built-in inference are:

 - New methods on the `Operator` trait which return shape and type inference
   rules for the operator.

 - Implementations of these methods for a subset of operators, sufficient to
   enable the new attention-related fusions used by eg. the 4-bit quantized
   Llama 3 model.

 - An `infer_graph` function in src/infer_shapes.rs which runs shape
   inference for all operators in a model and generates a map from value
   node ID to inferred shape and type metadata.

 - A `ModelOptions::enable_shape_inference` method which enables the
   experimental built-in shape inference pass.

 - Debug logging which is enabled by setting the `RTEN_INFER_SHAPES_DEBUG`
   env var to "1". This is similar to how the `RTEN_TIMING` env var works.

Part of https://github.com/robertknight/rten/issues/1091