Add rten-shape-inference crate with symbolic tensor types
Add a new crate which will contain code to support shape inference of ONNX
graphs. The initial content are types to represent tensors with symbolic shapes
and values. These can be used to represent values during symbolic evaluation of
a graph. For example an image model may have an input with symbolic shape
("batch", 3, "height", "width"). Symbolic evaluation of a `Shape` operator
applied to this input returns a vector `["batch", 3, "height", "width"]`.

The set of symbolic expressions initially supported is the minimum needed for
complete shape inference of the Llama 3 model with dynamic batch size and
sequence length.

Part of https://github.com/robertknight/rten/issues/1091