Add shape inference traits and common implementations
Add traits for shape inference and common implementations for unary, binary and
reduction ops.

The common implementations handle shape inference but ignore any symbolic values
in the tensor. This means that the symbolic output always contains just a shape
and any information about the values is lost.

This commit also updates the crate docmentation to illustrate how symbolic
shapes and values are created and manipulated during shape inference.