Fix confusing error message when ONNX model contains unsupported op
When loading an ONNX model with an unsupported operator, the error message
stated that the operator was "not enabled". This is due to some logic originally
written for the .rten loader which assumes that if the operator name is known,
the operator is supported. With ONNX models however, the operator name is known
even if the operator isn't supported.