Use shape inference results to replace more values with constants
Shape inference may determine that a node has a fixed constant output which
doesn't depend on model inputs. In that case the value node in the graph can be
replaced by a constant, avoiding the need to compute it on each inference. Since
ML models typically consist of many identical or at least similar layers, many
of these inferred constants will have the same value.

For the Llama 3 3B model, shape inference infers that 960 value nodes have
constant values, of which there are only 21 unique values. Replacing these 960
values with constants reduces the number of operators run on each generation
step from 2022 to 1740 (-14%). The reduction in op count is smaller than the
number of value-to-constant substitutions because some of these substitutions
were previously performed later in the optimization process by constant
propagation.

The impact on runtime is small, since the removed operations are cheap ops that
operate on small tensors. This nevertheless saves interpreter overhead over the
course of generation and replacing more values with constants can enable
additional optimizations elsewhere.