WIP - Use shape inference results to avoid recomputing equal values
Shape inference can determine sets of values that have equal symbolic values,
meaning that in a given graph execution, all values in the set will compute the
same concrete value. Only the first node from each set, in the topological
ordering of the graph, needs to be computed. Uses of the other nodes can be
replaced with the first.

Applying this to the Llama 3 3B model, this reduces the number of ops that need
to be run for each inference from 1740 to 1259 (-28%). These are cheap
operations manipulating tiny tensors, so the impact is small.