Add basic optimizer diagnostics infrastructure
Add a minimal diagnostic reporter for the graph optimizer, which is enabled by
setting the `RTEN_OPTIMIZER_DEBUG` env var to a truthy value.  Initially this
reports a single diagnostic, which is for fusions that are skipped because an
intermediate value in the fused subgraph is used outside the subgraph.

Part of https://github.com/robertknight/rten/issues/1110.