Add `/read {filename}` special command to Llama chat example
Support "/read {filename}" as a way to read text files into the context in the
Llama example. This makes the example more useful, but also shows a shortcoming
of the current implementation which is that prompt processing is slower than it
could be, and only usable with text files that are a few KB in size.