Add `info` level to optimizer diagnostics
Allow the optimizer diagnostics to be set to one of three levels: off, warn or
info. When set to "info", the optimizer reports fusions that were successfully
applied in addition to warnings about unsuccessful fusions.

The diagnostic level is configured by setting `RTEN_OPTIMIZER_DEBUG` to "warn"
or "info". If set to a truthy value ("y", "1" etc.) it behaves like "warn".

Part of https://github.com/robertknight/rten/issues/1110.