Output diagnostics when fusion preconditions are not met
Change the fusion traits to enable fusions to return a reason why a fusion was
not created. Use this to report info-level diagnostics when a fusion's basic
pattern matches a graph node, but a pre-condition (eg. on operator
attributes, inputs etc.) is not met.

This should make debugging failed optimizations a more efficient process.

Part of https://github.com/robertknight/rten/issues/1110