Improve shape inference for Slice op when input is a symbolic vector
When the input is a symbolic vector, make the result also a (sliced) symbolic
vector instead of just a shape. Implementing this would have involved copying
additional parts of `SliceRange` from rten-tensor into the rten-shape-inference
crate, so I bit the bullet and added rten-tensor as a dependency instead. This
could be avoided by splitting out the `SliceRange` type into its own mini-crate
in future.