Enable addition and subtraction to cancel
Enable addition and subtraction of the same symbol to cancel. This enables
simplifying `x + y - x` to `y` for example. Eliminating the subtraction makes it
possible to reason that the expression evaluates to a positive value, if all the
symbols are positive. This is useful for a common ONNX subgraph of the form
`Where(Equal(shape, -1), x, y)`. If we can reason that all elements of `shape`
are positive, then the output is equal to `y`.

The implementation works by:

 1. Changing canonicalization to rewrite subtraction as addition of
    negation (`x - y` => `x + (-x)`).
 2. Sorting terms by symbol name during re-association (`x + y - x` =>
    `x + (-x) + y`)
 3. Adding a step after re-association that removes adjacent terms which
    are negations of each other (`x + (-x) + y` => `y`).