Add `broadcast` and `min` symbolic expressions
Add symbolic expressions for the broadcast and min operations and use them to
improve shape inference for binary and Slice ops.

`broadcast(x, y)` is the function used to broadcast two dimension sizes in
various ONNX ops. It functions like `max` but implies more constraints on the
arguments: they must be positive and either equal or one of them must be 1.
`min` is useful to represent the output size in a Slice op if both input
dimension sizes are symbolic.

Shape inference for the bert-base model now infers the output shape as
`[batch_size, broadcast(sequence_length, min(sequence_length, 512)), 30522]`.
This implies an inconsistency if `sequence_length` exceeds 512, and indeed
inference fails if that is the case. I haven't decided how to handle these
constraints on symbolic dimensions uncovered through shape inference yet, but it
would be useful to expose to the user.