Add shape inference for convolution and pooling ops
Enable shape inference for:

 - AveragePool
 - Conv
 - GlobalAveragePool
 - GlobalMaxPool
 - MaxPool

Most of the complexity for these operators is in the inference of the output
spatial shapes, which uses the same logic.

Use of these ops currently leads to excessively complex shape expressions in
later ops. This is because division expressions are not simplified as well as
they could be.