Improve handling of unnamed dimensions when reading ONNX value info
If a value has a dimension without a size or name, generate a unique dummy name
for it and include the dimension in the shape metadata. Previously the
dimension was skipped, resulting in value metadata with the wrong rank. Missing
names can still limit model optimization because each anonymous dimension has to
be treated as being a distinct value.

This resolves an issue with running the Tiny YOLO model mentioned in
https://github.com/robertknight/rten/pull/1153 via the CLI.