Simplify repeated division by constants in symbolic expressions
Simplify `x / b / c => x / (b * c)` where possible. This kind of repeated
division often comes up when an input is fed through multiple convolution or
pooling layers. In this case we want output value shapes to be inferred as eg.
`width / 32` rather than `width / 2 / 2 / 2 / 2 / 2`. For ceiling division, a
similar simplification is possible but with different constraints on the
constants.