Add `Generator::clear_prompt` API
This is useful for clearing the pending prompt created by `next` or
`append_prompt`. A use case for this is to feed a long prompt into a model in
chunks in order to reduce memory usage. In this case, we want to remove the
token that is sampled and added to the prompt after each chunk, except for the
final chunk.