Add API to process prompt without generating or sampling logits
Add `Generator::process_prompt` API to process the current prompt and update the
KV-cache without generating output logits or sampling a token. When processing
a prompt, the logits output and token sampling is only needed for the last
position of the prompt. For other positions only the updated KV-cache is needed.
The logits matmul is usually expensive due to the large vocab size in many
models, so it is useful to have an API to skip it.

A caveat with using this API is that rten's `Graph` only internally caches a
single execution plan, so switching between generating KV-cache outputs only or
KV-cache + logits will force regeneration of the plan. This is typically cheap
(~1ms) compared to token generation however.

Some additional `Generator` methods for querying the current prompt and KV cache
length have also been added as these were useful in tests.