Support f16 models in ONNX loader by up-converting to f32
Make it possible to run models with f16 weights by converting FLOAT16
initializers to f32 when loading the model.

The implementation uses a copy of the f16 <-> f32 conversion functions from the
`half` crate. I didn't add that crate as a dependency yet because it brings in
additional deps with proc macros (zerocopy). The conversion does get vectorized,
but could be made more efficient by using the native x86/Arm functions for this.
It will probably make sense to do that via additions to rten-simd.