Add strict shape inference mode
Add a strict mode for shape inference. When enabled, models will fail to load if
shape inference is unsupported or is incomplete for any operators. This makes it
easier to check if shape inference encounters any problems that might limit
optimization.

In the CLI, the `--infer-shapes` switch has been changed to an option which
accepts "off", "on" or "strict" values.