Remove common factors when simplifying division, use in Reshape op
Implement removal of common factors when simplifying division expressions in
shape inference. This enables using symbolic division to update the remainder in
Reshape shape inference. This in turn avoids the need to use unknown values as
the remainder.

With this change, shape inference for the CLIP model now almost works.  The last
missing piece after this is supporting shape inference for ArgMin and ArgMax.

There is a change in this commit to avoid creating redundant `1 * x` and `0 + x`
terms when reassociating nested associative ops. These get removed later during
simplification, so the output doesn't change, but it removes some noise from
un-simplified expressions.