Make conversion of `to` attribute for Cast op match initializer loader
Make the conversion of ONNX dtypes to RTen's internally supported dtypes in
`onnx_registry.rs` match those used when converting initializers and value info
in `onnx_loader.rs`.

This enables loading models with `Cast` nodes where the target type is f64. As
with f64 -> f32 conversion of initializers, this may potentially affect
accuracy, but this has not been a problem with real models so far. See also
https://github.com/robertknight/rten/issues/1176.